+++ src/view/frontend/web/js/view/shipping.js
@@ -14,8 +14,8 @@ define(function () {
             setShippingInformation: function () {
                 if (typeof(ga) != "undefined") {
                     var urlToTrack = foomanGaBaseUrl + '/payment';
-                    if (foomanGaQuery.length > 0) {
-                        urlToTrack += '?' + foomanGaQuery
+                    if (location.search.length > 0) {
+                        urlToTrack += location.search
                     }
                     ga('set', 'page', urlToTrack);
                     ga('send', 'pageview');