+++ tests/integration/testsuite/Fooman/GoogleAnalyticsPlus/Block/JsTest.php
@@ -21,7 +21,6 @@ class JsTest extends \Magento\TestFramework\TestCase\AbstractController
         $this->dispatch('');
         $body = $this->getResponse()->getBody();
         $this->assertContains("var foomanGaBaseUrl = '';", $body);
-        $this->assertContains("var foomanGaQuery = '';", $body);
     }
 
     /**
@@ -34,6 +33,5 @@ class JsTest extends \Magento\TestFramework\TestCase\AbstractController
         $this->dispatch('/cms/index/index?param1=key1&param2');
         $body = $this->getResponse()->getBody();
         $this->assertContains("var foomanGaBaseUrl = '/cms';", $body);
-        $this->assertContains("var foomanGaQuery = 'param1=key1&param2';", $body);
     }
 }