+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+## [Unreleased]
+
 ## [104.0.1] 2018-12-02
 ### Changed
 - Reorganised unit tests