+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Change Log
+## [Unreleased]
+### Security
+- Fixed potential XSS due to wrong escape method used
 
 ## [2.0.5] 2017-06-05
 ### Changed